results/rebuttal_new_baselines/run-20231112-040610_CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-recipe-donnemartin-system-design-oop-eventplanner_0_20-runs_log/recipe/18/Reflexion/test_app.py
import pytest
from app import app, User

@pytest.fixture

def client():
	app.config['TESTING'] = True
	with app.test_client() as client:
		yield client


def test_create_user(client):
	response = client.post('/user', json={'id': 1, 'username': 'test', 'password': 'test'})
	assert response.status_code == 201
	assert response.get_json() == {'id': 1}


def test_get_user(client):
	client.post('/user', json={'id': 1, 'username': 'test', 'password': 'test'})
	response = client.get('/user/1')
	assert response.status_code == 200
	assert response.get_json() == {'id': 1, 'username': 'test'}


def test_get_user_not_found(client):
	response = client.get('/user/1')
	assert response.status_code == 404
	assert response.get_json() == {'error': 'User not found'}
